Execute surrogate minimal depth variable importance

This function determines the surrogate minimal depth of variables from a forest that is created by [getTreeranger()], [addLayer()] and [addSurrogates()] functions.

Usage

surrmindep(forest, s.l = count.surrogates(forest$trees)$s.l)

Arguments

forest

a list containing `allvariables` and `trees`. Allvariables is a vector of all variable names in the original data set (strings). Trees is a list of trees that was generated by [getTreeranger()], [addLayer()], and [addSurrogates()].

s.l

Number of average surrogate variables in the respective layers. (use [count.surrogates()] function to get it)

Value

List with the following components: * `depth`: mean surrogate minimal depth for each variable * `selected`: variables has been selected (1) or not (0), * `threshold`: the threshold that is used for the selection